Added main start line. Will implement lexer next.
Main start line has a gimmick with the username and the amount of times that the user has launched c-term. I'm not sure why I implemented this, or why it took so long for me to do so, but things happen, and in the name of the man Deltact himself, when you can do a cool gimmick, you should do that cool gimmick. Did it cost me 2 functions and some valuable memory? Yes. Do I care? Nah, not really.