Implementation of CHCommand and Help command
CHCommand is a command that also takes the command handler as an argument. It is a special type of command because it is usually meant to access something that wouldn't be accessed by normal commands. Normal commands are still in date. The help command provides a list of all commands and CHCommands and gives the descriptions, aliases and usages of each.