No longer using ArrayLists, SlowPrint command implementation
ArrayLists are now outdated and List<string>'s will be used in place as ArrayLists caused a specific issue with SlowPrint which caused me to think of updating the entire C-Term repository to use Lists instead. The SlowPrint commands features the implementation of DNP (see more in SlowPrint.cs). ArrayLists will still be used in specific places.