Implementation of specific command search in help command
I am so done with this shit, I just realized lists have a .Contain function to check if a specific variable exists in that List. I was using a for loop all this time. All good, I'll let the old code be, maybe revamp it later on to use .Contain instead.