All fs commands now use handler.currentDirectory
This is a change that is for the implementation of file browsing in c-term. Every filesystem command now uses CHCommand to be able to access handler. The next commit will involve making all CHCommands into the Command class but including the handler as a default argument. The commit after that will involve a settings command because the currentDirectory popup is a little too cluttery for the CUI + maybe I'd like to implement further changeable things into the future. The commit after that one will finalize c-term v2.0.0, allowing you to fully browse through whichever folders you are in like WinRAR and other apps and allowing you to modify them, create them and remove them within just a few keyboard taps. Note, the final commit might not be a single one because the task I am explaining is semi-difficult.